+++ node_transitions.go
@@ -35,6 +35,7 @@ func (n *Node) applyFollowerTransition(leaderID NodeID) {
 	n.transferTarget = ""
 	n.transferElapsed = 0
 	n.pendingConfigIndex = 0
+	n.capEntryPending = 0
 	n.termStartIndex = 0
 	n.leaderNopCommitted = false
 	// Reject all in-flight client proposals. If an uncommitted entry is later